Reorganize database tests
