+++ src/databaseTests/src/test/kotlin/org/vaccineimpact/api/databaseTests/touchstoneRepository/GetScenariosTests.kt
@@ -1,35 +1,21 @@
-package org.vaccineimpact.api.databaseTests
+package org.vaccineimpact.api.databaseTests.touchstoneRepository
 
-import org.assertj.core.api.Assertions.*
+import org.assertj.core.api.Assertions
 import org.junit.Test
-import org.vaccineimpact.api.app.errors.UnknownObjectError
 import org.vaccineimpact.api.app.filters.ScenarioFilterParameters
 import org.vaccineimpact.api.app.repositories.TouchstoneRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqScenarioRepository
-import org.vaccineimpact.api.app.repositories.jooq.JooqTouchstoneRepository
-import org.vaccineimpact.api.db.JooqContext
 import org.vaccineimpact.api.db.direct.*
 import org.vaccineimpact.api.models.*
 
-class TouchstoneTests : RepositoryTests<TouchstoneRepository>()
+class GetScenariosTests : TouchstoneRepositoryTests()
 {
-    val touchstoneName = "touchstone"
-    val touchstoneVersion = 1
-    val touchstoneId = "$touchstoneName-$touchstoneVersion"
-
-    override fun makeRepository(): TouchstoneRepository
-    {
-        val scenarioRepository = { JooqScenarioRepository() }
-        return JooqTouchstoneRepository(scenarioRepository)
-    }
-
     @Test
     fun `no scenarios are returned if touchstone is empty()`()
     {
         given {
             createTouchstoneAndScenarioDescriptions(it)
         } check {
-            assertThat(getScenarios(it)).isEmpty()
+            Assertions.assertThat(getScenarios(it)).isEmpty()
         }
     }
 
@@ -128,7 +114,7 @@ class TouchstoneTests : RepositoryTests<TouchstoneRepository>()
             createTouchstoneAndScenarioDescriptions(it)
             it.addScenarios(touchstoneId, "yf-1", "yf-2", "ms-1", "ms-2")
         } check {
-            assertThat(getScenarios(it, filter())).hasSize(4)
+            Assertions.assertThat(getScenarios(it, filter())).hasSize(4)
             checkResultAssumingNoCoverageSets(getScenarios(it, filter(scenarioId = "yf-1")), listOf(
                     Scenario("yf-1", "Yellow Fever 1", "YF", listOf(touchstoneId))
             ))
@@ -136,47 +122,7 @@ class TouchstoneTests : RepositoryTests<TouchstoneRepository>()
                     Scenario("ms-1", "Measles 1", "Measles", listOf(touchstoneId)),
                     Scenario("ms-2", "Measles 2", "Measles", listOf(touchstoneId))
             ))
-            assertThat(getScenarios(it, filter(scenarioId = "yf-1", disease = "Measles"))).isEmpty()
-        }
-    }
-
-    @Test
-    fun `getScenario throws exception if scenario doesn't exist`()
-    {
-        given {
-            createTouchstoneAndScenarioDescriptions(it)
-        } check {
-            assertThatThrownBy { it.getScenario(touchstoneId, "yf-1") }
-                    .isInstanceOf(UnknownObjectError::class.java)
-        }
-    }
-
-    @Test
-    fun `can get scenario with ordered coverage sets`()
-    {
-        val scenarioId = 1
-        val setA = 1
-        val setB = 2
-        val extraTouchstoneId = "extra-1"
-        given {
-            createTouchstoneAndScenarioDescriptions(it)
-            it.addTouchstone("extra", 1, addName = true)
-            it.addScenarioToTouchstone(touchstoneId, "yf-1", id = scenarioId)
-            it.addScenarioToTouchstone(touchstoneId, "yf-2", id = scenarioId + 1)
-            it.addScenarioToTouchstone(extraTouchstoneId, "yf-1", id = scenarioId + 2)
-            it.addCoverageSet(touchstoneId, "YF without", "YF", "without", "campaign", id = setA)
-            it.addCoverageSet(touchstoneId, "YF with", "YF", "with", "campaign", id = setB)
-            it.addCoverageSetToScenario(scenarioId, setB, 4)
-            it.addCoverageSetToScenario(scenarioId, setA, 0)
-        } check {
-            val result = it.getScenario(touchstoneId, "yf-1")
-            assertThat(result.scenario).isEqualTo(Scenario(
-                    "yf-1", "Yellow Fever 1", "YF", listOf(touchstoneId, extraTouchstoneId)
-            ))
-            assertThat(result.coverageSets).hasSameElementsAs(listOf(
-                    CoverageSet(setA, touchstoneId, "YF without", "YF", GAVISupportLevel.WITHOUT, ActivityType.CAMPAIGN),
-                    CoverageSet(setB, touchstoneId, "YF with", "YF", GAVISupportLevel.WITH, ActivityType.CAMPAIGN)
-            ))
+            Assertions.assertThat(getScenarios(it, filter(scenarioId = "yf-1", disease = "Measles"))).isEmpty()
         }
     }
 
@@ -190,22 +136,6 @@ class TouchstoneTests : RepositoryTests<TouchstoneRepository>()
         return it.scenarios(touchstoneId, filterParameters)
     }
 
-    private fun createTouchstoneAndScenarioDescriptions(it: JooqContext)
-    {
-        it.addTouchstone(touchstoneName, touchstoneVersion, addName = true, addStatus = true)
-        it.addDisease("YF", "Yellow Fever")
-        it.addDisease("Measles", "Measles")
-        it.addScenarioDescription("yf-1", "Yellow Fever 1", "YF")
-        it.addScenarioDescription("yf-2", "Yellow Fever 2", "YF")
-        it.addScenarioDescription("ms-1", "Measles 1", "Measles")
-        it.addScenarioDescription("ms-2", "Measles 2", "Measles")
-
-        it.addSupportLevels("none", "without", "with")
-        it.addActivityTypes("none", "routine", "campaign")
-        it.addVaccine("YF", "Yellow Fever")
-        it.addVaccine("Measles", "Measles")
-    }
-
     private fun checkResult(actuals: List<ScenarioAndCoverageSets?>, expecteds: List<ScenarioAndCoverageSets?>)
     {
         val paddedActuals = actuals + listOfNulls(expecteds.size - actuals.size)
@@ -214,16 +144,16 @@ class TouchstoneTests : RepositoryTests<TouchstoneRepository>()
         {
             if (expected == null)
             {
-                fail("Unexpected element: " + actual)
+                Assertions.fail("Unexpected element: " + actual)
             }
             else if (actual == null)
             {
-                fail("Missing expected element: " + expected)
+                Assertions.fail("Missing expected element: " + expected)
             }
             else
             {
-                assertThat(actual.scenario).isEqualTo(expected.scenario)
-                assertThat(actual.coverageSets).hasSameElementsAs(expected.coverageSets)
+                Assertions.assertThat(actual.scenario).isEqualTo(expected.scenario)
+                Assertions.assertThat(actual.coverageSets).hasSameElementsAs(expected.coverageSets)
             }
         }
     }
